chore(ci): update mac ci resource
Mac medium and large resources are being deprecated so updating this now, https://discuss.circleci.com/t/macos-resource-deprecation-update/46891. The new large resource is available as an m1, but the runtime cost is 4x the current large, https://circleci.com/product/features/resource-classes/#macos, so moving this back to medium for the time being.